requirements.txt
influxdb-client==1.45.0
python-telegram-bot==21.4
python-dotenv==1.0.1
